Refactored testing to add an integration test project for the functions. The intention is to start the local emulator, and call the functions to test actually hitting the Azure infrastructure from the function, using Azure DevTest Labs in the CI/CD process.
