+++ .gitignore
@@ -37,6 +37,7 @@ bld/
 # NUNIT
 *.VisualState.xml
 TestResult.xml
+*.runsettings
 
 # Build Results of an ATL Project
 [Dd]ebugPS/
