+++ Azure.DataPipelineTools.sln
@@ -12,7 +12,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		LICENSE = LICENSE
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataPipelineTools.Tests", "DataPipelineTools.Tests\DataPipelineTools.Tests.csproj", "{A2C01394-16F9-4783-9629-2857C400D6E6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataPipelineTools.Tests", "DataPipelineTools.Tests\DataPipelineTools.Tests.csproj", "{A2C01394-16F9-4783-9629-2857C400D6E6}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataPipelineTools.Functions.Tests", "DataPipelineTools.Functions.Tests\DataPipelineTools.Functions.Tests.csproj", "{C2448E27-7F4C-4E1E-BAF7-D5344F330073}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataPipelineTools.Tests.Common", "DataPipelineTools.Tests.Common\DataPipelineTools.Tests.Common.csproj", "{6C304777-0D1A-45A6-A5E0-5849022F319A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -32,6 +36,14 @@ Global
 		{A2C01394-16F9-4783-9629-2857C400D6E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A2C01394-16F9-4783-9629-2857C400D6E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A2C01394-16F9-4783-9629-2857C400D6E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C2448E27-7F4C-4E1E-BAF7-D5344F330073}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C2448E27-7F4C-4E1E-BAF7-D5344F330073}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2448E27-7F4C-4E1E-BAF7-D5344F330073}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C2448E27-7F4C-4E1E-BAF7-D5344F330073}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6C304777-0D1A-45A6-A5E0-5849022F319A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6C304777-0D1A-45A6-A5E0-5849022F319A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6C304777-0D1A-45A6-A5E0-5849022F319A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6C304777-0D1A-45A6-A5E0-5849022F319A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
