+++ DataPipelineTools.Functions.Tests/DataPipelineTools.Functions.Tests.csproj
@@ -1,0 +1,27 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+
+    <RootNamespace>DataPipelineTools.Functions.Tests</RootNamespace>
+
+    <AssemblyName>DataPipelineTools.Functions.Tests</AssemblyName>
+
+    <PackageId>DataPipelineTools.Functions.Tests</PackageId>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
+    <PackageReference Include="NUnit" Version="3.13.1" />
+    <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
+    <PackageReference Include="coverlet.collector" Version="3.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\DataPipelineTools.Functions\DataPipelineTools.Functions.csproj" />
+    <ProjectReference Include="..\DataPipelineTools.Tests.Common\DataPipelineTools.Tests.Common.csproj" />
+  </ItemGroup>
+
+</Project>
