+++ DataPipelineTools.Functions.Tests/example.runsettings
@@ -1,0 +1,65 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<RunSettings>
+  <!-- Parameters used by tests at run time -->
+  <TestRunParameters>
+    <!-- 
+    **********************************************************************************************************************************************
+    **** Integration Test Params                                                         
+    ****
+    **** The integration tests can't use the Azure Storage emulator as is does not support hierarchical storage accounts (ie ADLS). This means
+    **** to run the integration tests you need a storage account/container with the test file structure added to it, and a Azure Key Vault to
+    **** hold keys for testing connectivity.
+    **** 
+    **** The following credentials are required:
+    ****    - Functions App Service Principal with access to the ADLS container and an Azure Key Vault to retrieve secrets
+    ****    - Service Principal and its secret (both directly and via Key Vault) with access to the ADLS container
+    ****    - SAS token (both directly and via Key Vault) for the ADLS container
+    ****    - Storage Account Access Key (both directly and via Key Vault) for the ADLS container
+    ****
+    **** Note: Secrets in this file are examples from the Azure portal that have been reset since.
+    ****
+    **********************************************************************************************************************************************
+    -->
+    
+    <!-- Set whether to use the local function emulator or not -->
+    <Parameter name="UseFunctionsEmulator" value="true" />
+    
+    <!-- The URL for the functions app. Only required if the emulator flag (UseFunctionsEmulator) above is false. -->
+    <Parameter name="FunctionsAppUrl" value="functionsapp.azurewebsites.net" />
+    <!--Parameter name="FunctionsUrl" value="http://localhost:7071/api/HttpTrigger" /-->
+
+    <!-- The key to authorise calls to the functions app. Only required if the emulator flag (UseFunctionsEmulator) above is false. -->
+    <Parameter name="FunctionsAppKey" value="" />
+
+    <!-- The name of the storage account -->
+    <Parameter name="StorageAccountName" value="testadlsaccount" />
+
+    <!-- The name of the storage container -->
+    <Parameter name="StorageContainerName" value="test" />
+
+    <!-- The name of the key vault -->
+    <Parameter name="KeyVaultName" value="test_key_vault" />
+
+    <!-- The name of the Service Principal -->
+    <Parameter name="ServicePrincipalName" value="testSpn" />
+
+    <!-- The secret/password for the Service Principal -->
+    <Parameter name="ServicePrincipalSecret" value="" />
+
+    <!-- The name of the Azure Key Vault secret holding the secret/password for the Service Principal -->
+    <Parameter name="ServicePrincipalNameKeyName" value="" />
+
+    <!-- The SAS token. This needs to be escaped for XML when you copy it from the Azure portal (it should only be the & sign that is an issue) -->
+    <Parameter name="SasToken" value="sp=r&amp;st=2021-07-01T13:16:28Z&amp;se=2021-07-01T21:16:28Z&amp;spr=https&amp;sv=2020-08-04&amp;sr=c&amp;sig=lUXpUl0u6kV1hrnIyQtIztOWIGfvmO56Nred4qP5dK8%3D" />
+
+    <!-- The name of the Azure Key Vault secret holding the SAS token -->
+    <Parameter name="SasTokenKeyName" value="testSasToken" />
+
+    <!-- The Storage Account Access Key -->
+    <Parameter name="AccessKey" value="nc8LtiyiKbeJawBW1pAJtOi4p6Id6rUqx1LqiY6zUZAplFEu9cuuT9YRlv+McTTqZJKI6RRhJDNTdPJykKk5wA==" />
+
+    <!-- The name of the Azure Key Vault secret holding the Storage Account Access Key -->
+    <Parameter name="AccessKeyKeyName" value="testAccessKey" />
+
+  </TestRunParameters>
+</RunSettings>
